fix: Move www redirect logic from middleware.ts to proxy.ts
Next.js doesn't allow both middleware.ts and proxy.ts to exist.
Merged www redirect logic into existing proxy.ts file.
