--- middleware.ts
@@ -1,27 +1,0 @@
-import { NextResponse } from 'next/server';
-import type { NextRequest } from 'next/server';
-
-export function middleware(request: NextRequest) {
-  const url = request.nextUrl.clone();
-  const hostname = request.headers.get('host') || '';
-
-  // Force www in production to ensure consistent redirect URLs for auth
-  if (process.env.VERCEL_ENV === 'production' && hostname === 'ridesharetahoe.com') {
-    url.host = 'www.ridesharetahoe.com';
-    return NextResponse.redirect(url, 301);
-  }
-
-  return NextResponse.next();
-}
-
-export const config = {
-  matcher: [
-    /*
-     * Match all request paths except for the ones starting with:
-     * - _next/static (static files)
-     * - _next/image (image optimization files)
-     * - favicon.ico (favicon file)
-     */
-    '/((?!_next/static|_next/image|favicon.ico).*)',
-  ],
-};
